lines 8000-16000
