+++ src/compiler/checker/lines_0_8000.rs
@@ -687,153 +687,6 @@ impl TypeChecker {
     pub(crate) fn get_type_name_for_error_display(&self, type_: Rc<Type>) -> String {
         self.type_to_string(type_)
     }
-
-    pub(crate) fn get_name_of_symbol_as_written(
-        &self,
-        symbol: Rc<Symbol>,
-        context: Option<&NodeBuilderContext>,
-    ) -> String {
-        if let Some(declarations) = &*symbol.maybe_declarations() {
-            if !declarations.is_empty() {
-                let declaration = first_defined(declarations, |d, _| {
-                    if get_name_of_declaration(&**d).is_some() {
-                        Some(d)
-                    } else {
-                        None
-                    }
-                });
-                let name = if let Some(declaration) = declaration {
-                    get_name_of_declaration(&**declaration)
-                } else {
-                    None
-                };
-                if let Some(name) = name {
-                    return declaration_name_to_string(Some(&*name));
-                }
-            }
-        }
-        unimplemented!()
-    }
-
-    pub(crate) fn add_optionality(
-        &self,
-        type_: Rc<Type>,
-        is_property: Option<bool>,
-        is_optional: Option<bool>,
-    ) -> Rc<Type> {
-        let is_property = is_property.unwrap_or(false);
-        let is_optional = is_optional.unwrap_or(true);
-        if self.strict_null_checks && is_optional {
-            self.get_optional_type(type_, Some(is_property))
-        } else {
-            type_
-        }
-    }
-
-    pub(crate) fn get_type_for_variable_like_declaration(
-        &self,
-        declaration: &Node,
-    ) -> Option<Rc<Type>> {
-        let is_property =
-            is_property_declaration(declaration) || is_property_signature(declaration);
-        let is_optional = false;
-
-        let declared_type = self.try_get_type_from_effective_type_node(declaration);
-        if let Some(declared_type) = declared_type {
-            return Some(self.add_optionality(declared_type, Some(is_property), Some(is_optional)));
-        }
-        unimplemented!()
-    }
-
-    pub(crate) fn get_widened_type_for_variable_like_declaration(
-        &self,
-        declaration: &Node,
-    ) -> Rc<Type> {
-        self.widen_type_for_variable_like_declaration(
-            self.get_type_for_variable_like_declaration(declaration),
-            declaration,
-        )
-    }
-
-    pub(crate) fn widen_type_for_variable_like_declaration(
-        &self,
-        type_: Option<Rc<Type>>,
-        declaration: &Node,
-    ) -> Rc<Type> {
-        if let Some(type_) = type_ {
-            return self.get_widened_type(type_);
-        }
-        unimplemented!()
-    }
-
-    pub(crate) fn try_get_type_from_effective_type_node(
-        &self,
-        declaration: &Node, /*Declaration*/
-    ) -> Option<Rc<Type>> {
-        let type_node = get_effective_type_annotation_node(declaration);
-        type_node.map(|type_node| self.get_type_from_type_node(&*type_node))
-    }
-
-    pub(crate) fn get_type_of_variable_or_parameter_or_property(
-        &self,
-        symbol: &Symbol,
-    ) -> Rc<Type> {
-        // let links = self.get_symbol_links(symbol);
-        // if links.type_.is_none() {
-        self.get_type_of_variable_or_parameter_or_property_worker(symbol)
-        // }
-        // links.type.unwrap().clone()
-    }
-
-    pub(crate) fn get_type_of_variable_or_parameter_or_property_worker(
-        &self,
-        symbol: &Symbol,
-    ) -> Rc<Type> {
-        Debug_.assert_is_defined(&symbol.maybe_value_declaration(), None);
-        let declaration = symbol
-            .maybe_value_declaration()
-            .as_ref()
-            .unwrap()
-            .upgrade()
-            .unwrap();
-
-        let type_: Rc<Type>;
-        if false {
-            unimplemented!()
-        } else if is_property_assignment(&*declaration) {
-            type_ = self
-                .try_get_type_from_effective_type_node(&*declaration)
-                .unwrap_or_else(|| {
-                    self.check_property_assignment(match &*declaration {
-                        Node::PropertyAssignment(property_assignment) => property_assignment,
-                        _ => panic!("Expected PropertyAssignment"),
-                    })
-                });
-        } else if is_property_signature(&*declaration) || is_variable_declaration(&*declaration) {
-            type_ = self.get_widened_type_for_variable_like_declaration(&*declaration);
-        } else {
-            unimplemented!()
-        }
-
-        type_
-    }
-
-    pub(crate) fn get_type_of_symbol(&self, symbol: &Symbol) -> Rc<Type> {
-        if symbol
-            .flags()
-            .intersects(SymbolFlags::Variable | SymbolFlags::Property)
-        {
-            return self.get_type_of_variable_or_parameter_or_property(symbol);
-        }
-        unimplemented!()
-    }
-
-    pub(crate) fn get_non_missing_type_of_symbol(&self, symbol: Rc<Symbol>) -> Rc<Type> {
-        self.remove_missing_type(
-            self.get_type_of_symbol(&*symbol),
-            symbol.flags().intersects(SymbolFlags::Optional),
-        )
-    }
 }
 
 fn create_node_builder() -> NodeBuilder {
